+++ src/core/Config.cpp
@@ -92,6 +92,8 @@ constexpr const int
 	quicksaveSlots = 3,
 	bufferSize = 0,
 	quickLevelTransition = JumpToChangeLevel;
+const int enemyHealth = 1;
+const bool creatureBars = false;
 
 const CutsceneAudience cutscenes = CutscenesForTrigger;
 
@@ -277,6 +279,8 @@ constexpr const std::string_view
 	useAltRuneRecognition = "improved_rune_recognition",
 	improvedBowAim = "improved_bow_aim",
 	quickLevelTransition = "quick_level_transition",
+	enemyHealth = "enemy_health",
+	creatureBars = "creature_health_bars",
 	allowConsole = "allow_console";
 
 // Input key options
@@ -542,6 +546,8 @@ bool Config::save() {
 	writer.writeKey(Key::useAltRuneRecognition, input.useAltRuneRecognition);
 	writer.writeKey(Key::improvedBowAim, input.improvedBowAim);
 	writer.writeKey(Key::quickLevelTransition, int(input.quickLevelTransition));
+	writer.writeKey(Key::enemyHealth, input.enemyHealth);
+	writer.writeKey(Key::creatureBars, input.creatureBars);
 	if(input.allowConsole) {
 		// Only write this if true so that switching from release to debug builds enables the console
 		writer.writeKey(Key::allowConsole, input.allowConsole);
@@ -683,6 +689,8 @@ bool Config::init(const fs::path & file) {
 	input.improvedBowAim = reader.getKey(Section::Input, Key::improvedBowAim, Default::improvedBowAim);
 	int quickLevelTransition = reader.getKey(Section::Input, Key::quickLevelTransition, Default::quickLevelTransition);
 	input.quickLevelTransition = QuickLevelTransition(glm::clamp(quickLevelTransition, 0, 2));
+	input.enemyHealth = glm::clamp(reader.getKey(Section::Input, Key::enemyHealth, Default::enemyHealth), 1, 3);
+	input.creatureBars = reader.getKey(Section::Input, Key::creatureBars, Default::creatureBars);
 	input.allowConsole = reader.getKey(Section::Input, Key::allowConsole, Default::allowConsole);
 	
 	// Get action key settings